Refactor id67 RGB Matrix file structure
- changed structure of the whole id67 directory so that it is split into RGB Lightning and RGB Matrix variants.
- renamed files so that the variants actually compile
- added keyboard level readme explaining the variants
