+++ keyboards/id67/default_rgb/config.h
