+++ keyboards/id67/default_rgb/default_rgb.c
@@ -14,4 +14,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "id67.h"
+#include "default_rgb.h"
