+++ keyboards/id67/default_rgb/info.json
