+++ keyboards/id67/default_rgb/readme.md
@@ -8,22 +8,12 @@ A 65% hotswap keyboard from IDOBAO.
 * Hardware Supported: IDOBAO ID67
 * Hardware Availability: [IDOBAO](https://www.idobao.net/products/idobao-id67-65-hot-swappable-mechanical-keyboard-kit-1)
 
-## Variants
-
-Currently there are two variants for the id67:
-1. `rgb_default` which uses the `RGB Lightning` feature for the on-board LEDs.
-2. `rgb` which uses the more advanced `RGB Matrix` feature for the per-key and underglow (bottom of PCB) LEDs.
-
-Make examples for this keyboard (after setting up your build environment):
+Make example for this keyboard (after setting up your build environment):
 
     make id67/default_rgb:default
 
-    make id67/rgb:default
-
-Flashing examples for this keyboard:
+Flashing example for this keyboard:
 
     make id67/default_rgb:default:flash
-
-    make id67/rgb:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
