+++ keyboards/id67/default_rgb/rules.mk
