+++ keyboards/id67/rgb/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "ID67",
+    "keyboard_name": "ID67 RGB Matrix variant",
     "url": "",
     "maintainer": "qmk",
     "layout_aliases": {
