--- keyboards/id67/rgb/keymaps/via/rules.mk
@@ -1,2 +1,0 @@
-LTO_ENABLE = yes
-VIA_ENABLE = yes
