+++ keyboards/id67/rgb/readme.md
@@ -19,11 +19,11 @@ A 65% hotswap keyboard from IDOBAO.
 
 Make example for this keyboard (after setting up your build environment):
 
-    make id67:rgb:default
+    make id67/rgb:default
 
 Flashing example for this keyboard:
 
-    make id67:rgb:default:flash
+    make id67/rgb:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
