+++ keyboards/id67/rgb/rgb.c
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "id67.h"
+#include "rgb.h"
 
 led_config_t g_led_config = { {
     // Key Matrix to LED Index
